Final Touches on the landing page v2
